+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/BaseSellerProtocol.java
@@ -20,6 +20,7 @@ package bisq.core.trade.protocol.bisq_v1;
 import bisq.core.trade.model.bisq_v1.SellerTrade;
 import bisq.core.trade.model.bisq_v1.Trade;
 import bisq.core.trade.protocol.FluentProtocol;
+import bisq.core.trade.protocol.SellerProtocol;
 import bisq.core.trade.protocol.TradeMessage;
 import bisq.core.trade.protocol.TradeTaskRunner;
 import bisq.core.trade.protocol.bisq_v1.messages.CounterCurrencyTransferStartedMessage;
